[TTS] Fix TTS audio preprocessing bugs
