+++ scripts/dataset_processing/tts/audio_processing/preprocess_audio.py
@@ -128,7 +128,7 @@ def _process_entry(
     if audio_trimmer is not None:
         audio, start_i, end_i = audio_trimmer.trim_audio(audio=audio, sample_rate=sample_rate, audio_id=audio_path)
 
-    if output_sample_rate is not None:
+    if output_sample_rate:
         audio = librosa.resample(y=audio, orig_sr=sample_rate, target_sr=output_sample_rate)
         sample_rate = output_sample_rate
 
@@ -140,7 +140,7 @@ def _process_entry(
     original_duration = librosa.get_duration(filename=audio_path)
     output_duration = librosa.get_duration(filename=output_path)
 
-    entry["duration"] = output_duration
+    entry["duration"] = round(output_duration, 2)
 
     if os.path.isabs(audio_filepath):
         entry["audio_filepath"] = output_path
